+++ aio2ch/__init__.py
@@ -1,3 +1,5 @@
 __author__ = 'wkpn'
-__version__ = '1.4.3'
+__version__ = '1.4.3.1'
 __license__ = 'MIT'
+
+from .api import Api