+++ tests/test_api.py
@@ -1,4 +1,4 @@
-from aio2ch.api import Api
+from aio2ch import Api
 from aio2ch.objects import *
 from aio2ch.exceptions import *
 import pytest
@@ -19,7 +19,15 @@ def test_api_url_changed():
 
 @pytest.mark.asyncio
 async def test_get_boards():
-    status, boards = await api.get_boards()
+    boards = await api.get_boards()
+
+    assert len(boards) > 0
+    assert all(isinstance(board, Board) for board in boards)
+
+
+@pytest.mark.asyncio
+async def test_get_boards_with_status():
+    status, boards = await api.get_boards(return_status=True)
 
     assert status >= 200
     assert len(boards) > 0
@@ -27,8 +35,16 @@ async def test_get_boards():
 
 
 @pytest.mark.asyncio
-async def test_get_board_threads(board='test'):
-    status, threads = await api.get_board_threads(board=board)
+async def test_get_board_threads():
+    threads = await api.get_board_threads(board='test')
+
+    assert len(threads) > 0
+    assert all(isinstance(thread, Thread) for thread in threads)
+
+
+@pytest.mark.asyncio
+async def test_get_board_threads_with_status():
+    status, threads = await api.get_board_threads(board='test', return_status=True)
 
     assert status >= 200
     assert len(threads) > 0
@@ -37,7 +53,15 @@ async def test_get_board_threads(board='test'):
 
 @pytest.mark.asyncio
 async def test_get_top_board_threads():
-    status, threads = await api.get_top_board_threads(board='test', method='views')
+    threads = await api.get_top_board_threads(board='test', method='views')
+
+    assert len(threads) == 5
+    assert all(isinstance(thread, Thread) for thread in threads)
+
+
+@pytest.mark.asyncio
+async def test_get_top_board_threads_with_status():
+    status, threads = await api.get_top_board_threads(board='test', method='views', return_status=True)
 
     assert status >= 200
     assert len(threads) == 5
@@ -52,10 +76,20 @@ async def test_get_top_board_threads_wrong_sort_method():
 
 @pytest.mark.asyncio
 async def test_get_thread_posts():
-    _, threads = await api.get_board_threads(board='test')
+    threads = await api.get_board_threads(board='test')
     thread = threads[0]
 
-    status, posts = await api.get_thread_posts(thread)
+    posts = await api.get_thread_posts(thread)
+
+    assert(all(isinstance(post, Post) for post in posts))
+
+
+@pytest.mark.asyncio
+async def test_get_thread_posts_with_status():
+    _, threads = await api.get_board_threads(board='test', return_status=True)
+    thread = threads[0]
+
+    status, posts = await api.get_thread_posts(thread, return_status=True)
 
     assert status >= 200
     assert(all(isinstance(post, Post) for post in posts))
@@ -83,9 +117,30 @@ async def test_get_thread_media():
 
     test_thread = Thread(test_thread_data, test_board)
 
-    status, thread_media = await api.get_thread_media(test_thread)
+    thread_media = await api.get_thread_media(test_thread)
 
-    assert status >= 200
     assert len(thread_media) > 0
     assert(all(isinstance(file, File) for file in thread_media))
 
+
+@pytest.mark.asyncio
+async def test_get_thread_media_with_status():
+    test_thread_data = {
+        'comment': '',
+        'num': 30972,
+        'posts_count': '',
+        'score': '',
+        'subject': '',
+        'timestamp': '',
+        'views': ''
+    }
+
+    test_board = 'test'
+
+    test_thread = Thread(test_thread_data, test_board)
+
+    status, thread_media = await api.get_thread_media(test_thread, return_status=True)
+
+    assert status >= 200
+    assert len(thread_media) > 0
+    assert (all(isinstance(file, File) for file in thread_media))